rbd: fix offset overflow on Windows
The rbd-wnbd daemon overflows when writing passed 4GB as the
offset field incorrectly uses uint32_t (initial versions of the
rbd-wnbd proposal used block increments).

This change updates the IO offset type to use uint64_t.

(cherry picked from commit 784e0356cac3dbf4935ef5afe088d98e4a3577d8)

+++ src/tools/rbd_wnbd/wnbd_handler.h
@@ -128,8 +128,8 @@ private:
     WnbdRequestType req_type = WnbdReqTypeUnknown;
     uint64_t req_handle = 0;
     uint32_t err_code = 0;
-    uint32_t req_size;
-    uint32_t req_from;
+    size_t req_size;
+    uint64_t req_from;
     bufferlist data;
 
     IOContext()
